Refactor settings view to use MVP
