+++ app/src/main/res/layout/activity_settings.xml
@@ -4,7 +4,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context="org.wit.hillfort.activities.SettingsActivity">
+  tools:context="org.wit.hillfort.views.settings.SettingsView">
 
   <RelativeLayout
     android:layout_width="match_parent"
